Moving functions around and renaming modules.
